ID to URL mappings are now written to cache
After being calculated, mappings will be written to the cache (and if present in cache, loaded at script start). This is done so that we don't have to recalculate them every single time we run, given that the entries will never go stale (though new ones might be added)

The cache will wind up being refreshed whenever the mappings are recalculated. At time of writing the things that cause this are

	- Running "list"
	- Trying to print a snippet that doesn't currently exist in the mappings (whether because it doesn't exist, or because it's new)